Move CIMXmlReader
